+++ leeJungWoo/models/postDao.js
@@ -39,7 +39,7 @@ const getAllPost = async () => {
   }
 };
 
-const getAllPostById = async (id) => {
+const getAllPostByUserId = async (id) => {
   try {
     return await database.query(
       `SELECT 
@@ -123,7 +123,7 @@ const increasePostLikes = async (userId, postId) => {
     );
   } catch (err) {
     const error = new Error('INVALID_DATA_INPUT');
-    error.statusCode = 500;
+    error.statusCode = 400;
     throw error;
   }
 };
@@ -131,7 +131,7 @@ const increasePostLikes = async (userId, postId) => {
 module.exports = {
   createPost,
   getAllPost,
-  getAllPostById,
+  getAllPostByUserId,
   patchContent,
   getPostByPostId,
   deletePostByPostId,
