+++ leeJungWoo/router/index.js
@@ -1,9 +1,9 @@
 const express = require('express');
 const router = express.Router();
-const homeRouter = require('./home.js');
+const userRouter = require('./user.js');
 const postRouter = require('./post.js');
 
-router.use('/', homeRouter);
+router.use('/', userRouter);
 router.use('/posts', postRouter);
 
 module.exports = router;
